Force string-typed keys
When unmarshaling objects to map[string]interface{}, map values will be
of the type map[interface{}]interface{}. This is problematic when
attempting to marshal the object out to JSON, since encoding/json only
supports string keys.

This change is a patch to cause all child maps to have string-typed
keys.